refactor to reuse fragments
