+++ packages/nextjs-contentful-renderer/src/utils/functional.ts
@@ -1,0 +1,3 @@
+export const compose = (f, g) => (x) => f(g(x));
+export const flip = (fn) => (...args) => fn(...args.reverse());
+export const composeRight = flip(compose);
